requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.1
Flask==3.0.3
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.4
MarkupSafe==2.1.5
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
sounddevice==0.4.6
spark-ai-python==0.3.24
tenacity==8.3.0
typing_extensions==4.11.0
urllib3==2.2.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
yarl==1.9.4



Levenshtein~=0.25.1
beautifulsoup4~=4.12.3
selenium~=4.20.0
opencv-python~=4.9.0.80
python-docx~=1.1.2